+++ stories/components/notice-bubble/docs/examples/undo_action.tsx
@@ -8,7 +8,7 @@ const manager = new NoticeBubbleManager();
 const Demo = () => {
   const openButtonRef = React.useRef<HTMLButtonElement>(null);
   const handleClick = () => {
-    const { focus } = manager.add({
+    manager.add({
       children: (
         <>
           Link was moved to <Link href='#'>Cats from outer space group</Link>
@@ -19,13 +19,10 @@ const Demo = () => {
       duration: 0,
       onClose: () => {
         setTimeout(() => {
-          if (document.activeElement === document.body) {
-            openButtonRef.current?.focus();
-          }
-        }, 500);
+          openButtonRef.current?.focus();
+        }, 300);
       },
     });
-    focus();
   };
 
   return (